refactor(timeline): extract gesture router from timeline-item host
Moves the ~250-line pointer-gesture router (click/double-click/mouse-down/
smart-trim-start) into use-timeline-item-pointer-handlers.ts. Verbatim move:
handler bodies unchanged; host values passed as a typed params object with
callback types derived from the source hooks. Stable refs/setters added to
dep arrays (idiomatic, behavior-preserving). Prunes 13 orphaned imports.

timeline-item/index.tsx: 1611 -> 1310 lines (cumulative 2070 -> 1310, -37%).

Note: verbatim move, but exercises interaction-heavy routing; smoke-test
gestures in-app before merge.
